+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/):
 
 - Custom tab name ([#3666])
 - New `--in` for `search` action to set search directory ([#3696])
+- Multi-file spotter ([#3733])
 - New `hovered` condition specifying different icons for hovered files ([#3728])
 - Allow using `ps.sub()` in `init.lua` directly without a plugin ([#3638])
 - New `sort_fallback` option to control fallback sorting behavior ([#3077])
@@ -1672,3 +1673,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/):
 [#3708]: https://github.com/sxyazi/yazi/pull/3708
 [#3725]: https://github.com/sxyazi/yazi/pull/3725
 [#3728]: https://github.com/sxyazi/yazi/pull/3728
+[#3733]: https://github.com/sxyazi/yazi/pull/3733